refactor: apply PR review suggestions
