+++ packages/proxy/README.md
@@ -1,34 +1,32 @@
-# APIMatic Test Utilities Libary for JavaScript
+# APIMatic Proxy Libary for JavaScript
 
 > This library is currently in preview.
 
-Provides assertion utilities for testing api calls. It can be plugged in as dev dependency to any library.
+Provides utility functions to generate HTTP and HTTPS proxy agents based on configuration. It supports conditional behavior for both Node.js and browser environments.
 
-The exported helper functions include:
+The exported helper functions and interfaces include:
 
-1. **getStreamData**: Get streaming data from a given URL.
-2. **toBuffer**: Promise to create a Buffer instance from a NodeJS.ReadableStream or Blob.
-3. **expectHeadersToMatch**: Compare actual headers with expected headers, ignoring case sensitivity.
-4. **expectMatchingWithOptions**: Check whether the expected value is matching with the actual value, with the given comparison options.
+1. **createProxyAgents**: Creates and returns HTTP and HTTPS proxy agents using provided proxy settings. In browser environments, this function returns undefined with a warning, as proxy agents are not supported in browsers.
+2. **ProxySettings**: Defines the proxy configuration, including a required address and optional port and authentication credentials.
 
-This library is used as a dev-dependency by JavaScript SDKs generated by the [APIMatic Code Generator](http://www.apimatic.io).
+This library is used by JavaScript SDKs generated by the [APIMatic Code Generator](http://www.apimatic.io).
 
 ## Builds
 
 The following environments are supported:
 
-1. Node.js v10+
+1. Node.js v14+ and v16+
 1. Bundlers like Rollup or Webpack
 1. Web browsers
 
 To support multiple environments, we export various builds:
 
-| Environment | Usage |
-| --- | --- |
-| Common.js | Import like this: `require('@apimatic/test-utilities')`. |
-| ES Module | Import like this: `import { /* your imports */ } from '@apimatic/test-utilities'`. |
-| Browsers | *Use script: `https://unpkg.com/@apimatic/test-utilities@VERSION/umd/schema.js` |
-| Modern Browsers (supports ESM and uses modern JS) | *Use script: `https://unpkg.com/@apimatic/test-utilities@VERSION/umd/schema.esm.js` |
+| Environment | Usage                                                                      |
+| --- |----------------------------------------------------------------------------|
+| Common.js | Import like this: `require('@apimatic/proxy')`.                            |
+| ES Module | Import like this: `import { /* your imports */ } from '@apimatic/proxy'`.  |
+| Browsers | *Use script: `https://unpkg.com/@apimatic/proxy@VERSION/umd/schema.js`     |
+| Modern Browsers (supports ESM and uses modern JS) | *Use script: `https://unpkg.com/@apimatic/proxy@VERSION/umd/schema.esm.js` |
 
 _* Don't forget to replace VERSION with the version number._
