+++ packages/proxy/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@apimatic/proxy",
-  "version": "0.0.1",
-  "description": "configure proxy agent",
+  "version": "0.1.0",
+  "description": "provides utilities to route requests through a configurable proxy",
   "main": "lib/index.js",
   "module": "lib/index.js",
   "types": "lib/index.d.ts",
