+++ packages/proxy/src/index.ts
@@ -1,2 +1,15 @@
-export { configureProxyAgent } from './proxyAgent';
-export { ProxySettings } from './proxySettings';
+export { createProxyAgents } from './proxyAgent';
+
+export interface ProxyAgents {
+  httpAgent: any;
+  httpsAgent: any;
+}
+
+export interface ProxySettings {
+  address: string;
+  port?: number;
+  auth?: {
+    username: string;
+    password: string;
+  };
+}
