+++ sonar-project.properties
@@ -2,10 +2,11 @@ sonar.projectKey=apimatic_apimatic-js-runtime
 sonar.organization=apimatic
 sonar.projectName=APIMatic JS Runtime
 sonar.host.url=https://sonarcloud.io
+sonar.sourceEncoding=UTF-8
 
 sonar.sources=packages
-sonar.tests=packages
+sonar.exclusions=**/*config.js
 
+sonar.tests=packages
 sonar.test.inclusions=**/test/**/*.test.ts
 sonar.javascript.lcov.reportPaths=packages/**/coverage/lcov.info
-sonar.sourceEncoding=UTF-8
